requirements.txt
opencv-python-headless
numpy
scikit-image
pillow
gradio
